Cleaned up the backend.
