+++ README.md
@@ -1,6 +1,13 @@
 # mock-ui
 A mock ui done in VueJS. 
 
+## Development 
+`npm ci`
+
+`npm run server`
+
+`npm run dev`
+
 ## Deploy to Test
 Now possible thanks to Github Actions! 
 Nothing has to be done. Fully automated deployment on master push.
