+++ server/engines/middleware.js
@@ -1,0 +1,9 @@
+const authMiddleware = (req, res, next) => {
+    if (!req.isAuthenticated()) {
+        res.status(401).send('You are not authenticated')
+    } else {
+        return next()
+    }
+};
+
+exports.authMiddleware = authMiddleware;
