+++ server/resource-access/users-static.js
@@ -1,0 +1,14 @@
+module.exports = [
+    {
+        id: 1,
+        name: "Jude",
+        email: "user@email.com",
+        password: "password"
+    },
+    {
+        id: 2,
+        name: "Emma",
+        email: "emma@email.com",
+        password: "password2"
+    },
+];
